pyproject.toml
[tool.poetry]
name = "myapp"
version = "0.1.0"
description = ""
authors = ["DucHoang-UIT <hoangduc.uit.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
notebook = "6.5.6"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.4.6"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.7"
psycopg2 = "^2.9.9"
python-decouple = "^3.8"
langchain = "^0.3.1"
langchain-openai = "^0.2.1"
rest-framework-simplejwt = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
